Cleaning and bugfixes...
git-svn-id: file:///Library/WebStart/SVN/NutShell@248 6e767ac9-f96a-0410-b219-bafc3452b216

+++ Frameworks/NutShell/trunk/classes/NLURLDownloader.m
@@ -173,7 +173,7 @@
     [ self dispatchEvent: @"ResponseReceived" ];
 }
 
-- ( void )download: ( NSURLDownload * )downloadObject didReceiveDataOfLength: ( unsigned int)length 
+- ( void )download: ( NSURLDownload * )downloadObject didReceiveDataOfLength: ( NSUInteger )length 
 {
     ( void )downloadObject;
